Crate constellation_traits

The interface to the Constellation, which prevents other crates from depending directly on the constellation crate itself. In addition to all messages to the Constellation, this crate is responsible for defining types that cross the process boundary from the embedding/rendering layer all the way to script, thus it should have very minimal dependencies on other parts of Servo.

Modules

from_script_message 🔒

Messages send from the ScriptThread to the Constellation.

structured_data 🔒

This module contains implementations of structured data as described in https://html.spec.whatwg.org/multipage/#safe-passing-of-structured-data

Structs

AuxiliaryWebViewCreationRequest

Specifies the information required to load an auxiliary browsing context.

AuxiliaryWebViewCreationResponse

Constellation’s response to auxiliary browsing context creation requests.

BlobImpl

The data backing a DOM Blob.

BroadcastChannelMsg

Message for communication between the constellation and a global managing broadcast channels.

DOMMessage

Message that gets passed to service worker scope on postMessage

DomException

A serializable version of the DOMException interface.

DomMatrix

A serializable version of the DOMMatrix/DOMMatrixReadOnly interface.

DomPoint

A serializable version of the DOMPoint/DOMPointReadOnly interface.

DomQuad

A serializable version of the DOMQuad interface.

DomRect

A serializable version of the DOMRect/DOMRectReadOnly interface.

FileBlob

File-based blob

IFrameLoadInfo

Specifies the information required to load an iframe.

IFrameLoadInfoWithData

Specifies the information required to load a URL in an iframe.

IFrameSizeMsg

An iframe sizing operation.

Job

https://w3c.github.io/ServiceWorker/#dfn-job

LoadData

can be passed to LoadUrl to load a page with GET/POST parameters or headers

MessagePortImpl

The data and logic backing the DOM managed MessagePort.

PortMessageTask

A task on the https://html.spec.whatwg.org/multipage/#port-message-queue

PortTransferInfo

The information needed by a global to process the transfer of a port.

SWManagerSenders

Channels to allow service worker manager to communicate with constellation and resource thread

ScopeThings

Entities required to spawn service workers

ScriptToConstellationChan

A Script to Constellation channel.

SerializableImageBitmap

A serializable version of the ImageBitmap interface.

SerializableImageData

SerializableIter

An iterator over the variants of Serializable

SerializableQuotaExceededError

A serializable version of the QuotaExceededError interface.

StructuredSerializedData

A data-holder for serialized data and transferred objects. https://html.spec.whatwg.org/multipage/#structuredserializewithtransfer

TransferableOffscreenCanvas

A struct supporting the transfer of OffscreenCanvas, which loosely corresponds to the dataHolder in https://html.spec.whatwg.org/multipage/#the-offscreencanvas-interface:offscreencanvas-16

TransferrableIter

An iterator over the variants of Transferrable

TransformStreamData

WorkerGlobalScopeInit

Resources required by workerglobalscopes

WorkerScriptLoadOrigin

Common entities representing a network load origin

Enums

BlobData

Different backends of Blob

DocumentState

Used to determine if a script has any pending asynchronous activity.

EmbedderToConstellationMessage

Messages to the Constellation from the embedding layer, whether from ServoRenderer or from libservo itself.

JobError

The kind of error the job promise should be rejected with.

JobResult

Messages sent from Job algorithms steps running in the SW manager, in order to resolve or reject the job promise.

JobResultValue

Jobs are resolved with the help of various values.

JobType

https://w3c.github.io/ServiceWorker/#dfn-job-type

JsEvalResult

The result of evaluating a javascript scheme url.

KeyboardScroll

An enum that describe a type of keyboard scroll.

LoadOrigin

The origin where a given load was initiated. Useful for origin checks, for example before evaluation a JS URL.

LogEntry

A log entry reported to the constellation We don’t report all log entries, just serious ones. We need a separate type for this because LogLevel isn’t serializable.

MessagePortMsg

Messages for communication between the constellation and a global managing ports.

NavigationHistoryBehavior

https://html.spec.whatwg.org/multipage/#navigation-supporting-concepts:navigationhistorybehavior

PaintMetricEvent

A description of a paint metric that is sent from the Servo renderer to the constellation.

SWManagerMsg

Messages outgoing from the Service Worker Manager thread to constellation

ScreenshotReadinessResponse

ScriptToConstellationMessage

Messages from the script to the constellation.

Serializable

All the DOM interfaces that can be serialized.

ServiceWorkerMsg

Messages sent to Service Worker Manager thread

Transferrable

All the DOM interfaces that can be transferred.

TraversalDirection

The direction of a history traversal

UserContentManagerAction

WindowSizeType

The type of window size change.

Traits

ServiceWorkerManagerFactory

This trait allows creating a ServiceWorkerManager without depending on the script crate.

Type Aliases

ScriptToConstellationSender